Extract ValidationRuleType enum
