+++ ddp-workspace/projects/ddp-sdk/src/lib/services/activity/validationRuleType.ts
@@ -1,0 +1,14 @@
+export enum ValidationRuleType {
+    Required = 'REQUIRED',
+    Complete = 'COMPLETE',
+    Length = 'LENGTH',
+    Regex = 'REGEX',
+    NumOptionsSelected = 'NUM_OPTIONS_SELECTED',
+    YearRequired = 'YEAR_REQUIRED',
+    MonthRequired = 'MONTH_REQUIRED',
+    DayRequired = 'DAY_REQUIRED',
+    DateRange = 'DATE_RANGE',
+    AgeRange = 'AGE_RANGE',
+    IntRange = 'INT_RANGE',
+    Unique = 'UNIQUE'
+}
